requirements.txt
Django~=2.2.4
django-environ==0.4.5
psycopg2==2.7.4 --no-binary psycopg2
djangorestframework==3.9.2
